milestone2/requirements.txt
# Core dependencies
pandas==2.0.3
numpy==1.24.3
chromadb==0.4.15
openai==1.3.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# ML and NLP
sentence-transformers==2.2.2
scikit-learn==1.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Database and storage
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data processing
openpyxl==3.1.2
xlrd==2.0.1

# Web and API
requests==2.31.0
httpx==0.25.1
aiohttp==3.9.0

# Additional utilities
python-multipart==0.0.6
email-validator==2.1.0
